requirements.txt
boto3==1.28.84
botocore==1.31.84
cffi==1.16.0
cryptography==41.0.5
jmespath==1.0.1
numpy==1.26.2
pandas==2.1.3
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
